NC-45 setup drwaer into NavItems.jsx
