+++ src/components/Navbar/Navbar.jsx
@@ -3,70 +3,35 @@
 /* eslint-disable @next/next/no-img-element */
 // Do not touch my Navbar. declared by the author - Zarif
 import { useForm } from "react-hook-form";
-import React, { useContext, useEffect, useRef, useState } from "react";
-import { GoHome } from "react-icons/go";
-import { IoNotificationsOutline, IoSettingsOutline } from "react-icons/io5";
-import { LuLogOut } from "react-icons/lu";
-import { RxCross2 } from "react-icons/rx";
-// import 'animate.css';
+import React, { useEffect, useRef, useState } from "react";
 import logo from "../../../public/swarm.png";
 import fakeUserData from "./fakeUsers.json";
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+import { BsExclamationCircle, BsSearch } from "react-icons/bs";
 import {
-  AiOutlineArrowRight,
-  AiOutlineProfile,
-  AiOutlineUser,
-} from "react-icons/ai";
-import { BsBookmarkCheck, BsExclamationCircle, BsSearch } from "react-icons/bs";
-import {
-  HiOutlineChatAlt2,
   HiOutlinePaperAirplane,
-  HiOutlineUserGroup,
 } from "react-icons/hi";
 // import component 👇
 import Image from "next/image";
-import Drawer from "react-modern-drawer";
 import "react-modern-drawer/dist/index.css";
 import { useRouter } from "next/navigation";
-
 import Link from "next/link";
-import ThemeButton from "./ThemeButton";
-import { TbLayoutDashboard } from "react-icons/tb";
 import SearchSection from "./SearchSection";
-import NavFeedback from "./NavFeedback";
-import AuthContext from "@/context/AuthContext";
-import Swal from "sweetalert2";
-import { RiDeleteBin5Line } from "react-icons/ri";
-import { LoggedInUser } from "@/hooks/loggedInUser";
+import NavItems from "./NavItems";
+import useAuth from "@/hooks/useAuth";
 
 
 const Navbar = () => {
-  const loggedInUer = LoggedInUser();
-  console.log(loggedInUer);
-  console.log("Log user is ", loggedInUer);
+  // const loggedInUer = LoggedInUser();
+  // console.log(loggedInUer);
+  // console.log("Log user is ", loggedInUer);
   const [success, setError] = useState([]);
   
-  const { user, logout } = useContext(AuthContext);
+  const { user, logout } = useAuth();
   console.log(user);
-  const handleLogOut = () => {
-    logout()
-      .then((data) => {
-        // console.log(data)
-        Swal.fire({
-          icon: "error",
-          title: "Caution",
-          text: "You logged Out!",
-          footer:
-            '<p></p><u><a href="/auth/signin" >Login</u></a> for having access</p>',
-        });
-      })
-      .catch((err) => {
-        console.log(err)
-      });
-  };
+  
 
-  // console.log(user);
   const {
     register,
     handleSubmit,
@@ -117,8 +82,6 @@ const Navbar = () => {
     }
    
   };
-
-  const router = useRouter();
   useEffect(() => {
     const handleClickOutside = (event) => {
       if (navbarRef.current && !navbarRef.current.contains(event.target)) {
@@ -138,7 +101,7 @@ const Navbar = () => {
   const navbarRef = useRef(null);
 
   const [searchActive, setSearchActive] = useState(false);
-  const [isOpen, setIsOpen] = React.useState(false);
+ 
   useEffect(() => {
     if (searchQuery.trim() === "") {
       setSearchResults([]);
@@ -163,144 +126,13 @@ const Navbar = () => {
     setSearchResults(filteredUsers);
   };
 
-  const toggleDrawer = () => {
-    setIsOpen((prevState) => !prevState);
-  };
+ 
   const handleKeyPress = (event) => {
     if (event.key === "Enter") {
       // Enter key was pressed, trigger the search button click event
       handleSearch();
     }
   };
-  const navItems = (
-    <>
-      {user ? (
-        <>
-          <li>
-            <Link href="/">
-              <p data-tip="Home" className="flex items-center tooltip-bottom tooltip  hover:scale-125 hover:translate-x-1 transform transition-transform ">
-                <GoHome className="text-2xl lg:text-2xl " />
-              </p>
-            </Link>
-          </li>
-        
-
-          <li>
-          {/* <li>
-            <a className="hover:bg-transparent hover:scale-125 transform transition-transform ">
-              <HiOutlineUserGroup className="text-2xl lg:text-2xl hover:scale-125 transform transition-transform  " />
-            </a>
-          </li>
-
-          <li> */}
-            <a data-tip="Message" className="hover:bg-transparent tooltip-bottom tooltip hover:scale-125 transform transition-transform">
-              <HiOutlineChatAlt2
-                onClick={() => router.push("/messages")}
-                className="text-2xl lg:text-2xl hover:scale-125 transform transition-transform "
-              />
-            </a>
-          </li>
-          {/* <li className="hidden lg:relative lg:block">
-            <a className="indicator hover:bg-transparent  hover:scale-125 transform transition-transform ">
-              <span className="lg:absolute left-3 top-2 indicator-item badge text-white bg-primary-color  font-bold px-2">
-                5
-              </span>
-              <IoNotificationsOutline className="rotate-45 text-xl lg:text-2xl hover:scale-125 transform transition-transform" />
-            </a>
-          </li> */}
-            <li>
-            <Link href={"/recycle"} data-tip="Recycle bin" className="hover:bg-transparent tooltip-bottom tooltip  hover:scale-125 transform transition-transform ">
-              <RiDeleteBin5Line className="text-2xl lg:text-2xl hover:scale-125 transform transition-transform  " />
-            </Link>
-          </li>
-          <li>
-            <a className=" hover:bg-transparent ">
-              <Image
-              alt="User image"
-                src={user?.photoURL}
-                width={32}
-                height={32}
-                className="h-8 w-8 rounded-full  hover:scale-125  transform transition-transform"
-                onClick={toggleDrawer}
-              ></Image>
-              <></>
-
-              <Drawer
-                open={isOpen}
-                onClose={toggleDrawer}
-                direction="right"
-                className="bla bla bla "
-              >
-                <div className="dark:bg-black-bg-primary  min-h-screen">
-                  <div className=" lg:text-lg flex flex-col gap-5 w-4/5 mx-auto pt-12   rounded-xl ">
-                    <p className=" ">
-                      <AiOutlineUser className="inline" />{" "}
-                      <span>{user.displayName}</span>
-                    </p>
-                    <hr className="text-primary-color" />
-                    <ThemeButton />
-                    <Link href="/Profile">
-                      <p className="  flex items-center group   hover:ml-2 transition-all">
-                        <AiOutlineProfile
-                          size={28}
-                          className="inline mr-2 rounded-full "
-                        />
-                        Profile
-                        {/* profile er vitore change password */}
-                        <AiOutlineArrowRight className="  ml-2 opacity-0 group-hover:opacity-100 inline" />
-                      </p>
-                    </Link>
-                    {/* <p className=" flex items-center group  hover:ml-2 transition-all">
-                      <IoSettingsOutline size={28} className="inline mr-2" />
-                      Settings
-                      <AiOutlineArrowRight className=" ml-2 opacity-0 group-hover:opacity-100 inline" />
-                    </p> */}
-                   <Link href="/bookmark"> <p className=" flex items-center group  hover:ml-2 transition-all">
-                      <BsBookmarkCheck size={28} className="inline mr-2" />
-                      Bookmarks
-                      <AiOutlineArrowRight className=" ml-2 opacity-0 group-hover:opacity-100 inline" />
-                    </p></Link>
-                    <NavFeedback />
-                   <Link href="/dashboard"> <p className=" flex items-center group  hover:ml-2 transition-all">
-                      <TbLayoutDashboard size={28} className="inline mr-2" />
-                      Dashboard
-                      <AiOutlineArrowRight className=" ml-2 opacity-0 group-hover:opacity-100 inline" />
-                    </p></Link>
-                    {/*  */}
-
-                   
-
-                    {/*  */}
-                    <p
-                      onClick={handleLogOut}
-                      className=" flex items-center group  hover:ml-2 transition-all"
-                    >
-                      <LuLogOut size={28} className="inline mr-2" />
-                      Log Out
-                      <AiOutlineArrowRight className=" ml-2 opacity-0 group-hover:opacity-100 inline" />
-                    </p>
-
-                    <div className="divider"></div>
-                  </div>
-                </div>
-              </Drawer>
-            </a>
-          </li>
-        </>
-      ) : (
-        <>
-          <div>
-            <Link
-              className="text-white hover:text-primary-color hover:bg-white bg-primary-color  py-3 shadow-sm hover:shadow-primary-color font-semibold px-3 rounded-xl"
-              href="/auth/signin"
-            >
-              Sign In
-            </Link>
-          </div>
-        </>
-      )}
-    </>
-  );
 
   return (
     <div className=" shadow-md shadow-slate-200 mt-3 lg:mt-0 z-50">
@@ -308,9 +140,6 @@ const Navbar = () => {
       <div className="my-container fixed  lg:glass bg-white z-50 mr-auto left-0 shadow-md shadow-slate-300 w-full lg:navbar myNav  lg:pb-0  items-center ">
         <div className="flex lg:flex lg:gap-64 items-center content-center z-50 w-[100px] mx-auto">
           <div className="navbar-start group ">
-            {/* responsive dropdown */}
-
-            {/* responsive dropdown */}
             <div className=" logo normal-case text-xl group-hover:animate-pulse lg:text-3xl hidden lg:block lg:mb-5">
               {" "}
               <Link href={"/"}>
@@ -320,17 +149,12 @@ const Navbar = () => {
             </div>
           </div>
         </div>
-
-
-
-
-        {/* search box */}
         <div className="navbar-end hidden lg:flex mx-auto ">
           <div className="hidden lg:block mr-12   mx-auto relative">
             <SearchSection />
           </div>
           <ul className="menu menu-horizontal px-1 ">
-            {navItems}
+           <NavItems/>
             {/* <AiOutlineUser  className='text-xl  lg:text-2xl'/> */}
           </ul>
         </div>
@@ -338,7 +162,7 @@ const Navbar = () => {
       {/* </Container> */}
       <ul className=" py-5 z-20 px-1 lg:hidden flex justify-center items-end absolute bottom-8 w-full ">
         <div className="fixed  bg-slate-200 bottom-0 bg-white  dark:bg-black py-4 px-5  w-11/12 flex gap-14 items-center justify-center content-center">
-          {navItems}
+          <NavItems/>
         </div>
         {/* <AiOutlineUser className='text-xl lg:text-2xl' /> */}
       </ul>
